Initial Terraform commit
When creating the organization, the bootstrap.sh script can be used to
setup the base for Terraform to execute:
- a bucket to store the state
- an IAM application to access the bucket
- secrets in the GitHub actions

The GitHub actions execute without the repository admin permission,
which prevents Terraform from reading the secrets inside a GitHub
action.

Since the secrets do not change often, let’s define them with the
bootstrap and rotate them without Terraform.