feat: Add a bucket for the institutional website
Sites-faciles needs a place to put the files uploaded to the Wagtail
CMS. Scalingo does not offer an S3-like tool, use Scaleway instead.

The CMS should handle serving files from the app using presigned URLs,
so no configuration is needed besides a private ACL.

The user and API key are managed manually for now, we will do better
when the scalingo app is managed by Terraform and we have a secret
manager instance to store that API key.

Use externally managed application to access the bucket

Until we have a secret manager and configure the scalingo app with
Terraform.