Add a do_not_use project to Scaleway
Previously, the default project was incorrectly used to create resources
with terraform. The old default project was renamed to
iac-gip-inclusion. Now, we’re creating a new project to use by default
in the Scaleway console, that should not hold resources. Resources
should always be created in a specific project (either for the codebase
that use it, or iac-gip-inclusion).