Remove the do-not-use project
We won’t be able to make do-not-use the default project, and the
Scaleway support confirmed the DNS root zone must live in the default
project (with the same UUID as the organization). So the default project
will necessarily hold some resources.

Since we can’t organize common resources in a dedicated project, the
best we can do is to document that the default project should not hold
resources, and monitor it.